requirements.txt
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.3
cryptography==3.1.1
decorator==4.4.2
pycodestyle==2.6.0
pycparser==2.20
python-telegram-bot==12.8
six==1.15.0
toml==0.10.1
tornado==6.0.4
